Use stof functions to parse hex strings as floating point params
This commit changes the Param.cc parsing of floating point values to always
use the stof family of functions to parse parameters given as hex strings
(e.g. strings starting with 0x or 0X). It removes the special case handling
for hex strings in the parser, and processes all strings using the same logic.
The isHex check is now used to specify the correct numeric base for the integer
conversion functions.

This commit also removes the boost special case in the HexFloat/HexDouble
Param.cc unit tests. Comments around the special case indicated that
boost::lexical_cast is able to handle parsing floats from hex values in boost
versions 1.58.0 and higher.  This appears not to be the case based on
issue #108. The stof family does handle these hex strings properly, so the
revised tests always expect successful parsing. Tests were also added to parse
the hex value 0X2A with a capital X and an arbitrary numerical value.

Tested with gcc 5.3.1 and boost 1.58.0 on Fedora 23.