Fix for compiling with Visual Studio and vcpkg
This commit contains two fixes:

* The first one is to avoid the hardcoding
of the location of the sdformat library (useful to avoid
in multi-configuration CMake generators such as Visual Studio).

* The second one modifies the logic of copying boost dll in the test
directory to support the installation layout of boost when installed
using vcpkg ( https://github.com/Microsoft/vcpkg ). The logic is still
quite hardcoded, but at least now it does not fail to compile if the files
are not found in the assumed location.