Significantly simplify Converter::Remove.
In particular, it has no reason to use boost:regex to split
on ::.  It should act a lot more like Add, where it just
takes a specific element or attribute to remove, and does
just that.  As an added bonus, we also add the ability to
remove attributes with this commit.