Fix parsing of URDF with an empty origin.
Previously, attempting to parse a URDF file with an empty
origin attribute of xyz or rpy would lead to a crash (attempting
to initialize a std::string with a NULL).  The new code detects
this situation and just adds an origin of 0,0,0,0,0,0 for an
empty origin.  Tests for this are also added.