Write a couple of tests for the stod conversions in parser_urdf.cc
In the first test, we iterate over a map that contains a list
of elements that we want to test for.  For each element, we
parse the XML and then check to make sure that all of the
elements that we expect to be there are actually there.  We
also check that the final number represented is what we expect.
The second test just checks that an invalid number throws a
std::invalid_argument like we expect.