Slightly improve performance in GetElement/GetAny.
Previously, we were calling HasElement() followed directly
by GetElement(), meaning we were iterating over the list
of Elements twice.  Instead, just call GetElement(); if it
returns ElementPtr() (i.e. nullptr), then we know that it
wasn't found and we can try our other strategies.  This
mostly matters if we have a large number of Elements, but
there is no reason to do the work twice.